Refactored, updated examples, appeased Sphinx.
