+++ README.rst
@@ -19,8 +19,6 @@ This library is compatible with standard Character LCDs such as:
 * `Adafruit RGB backlight negative LCD 16x2 <https://www.adafruit.com/product/399>`_
 * `Adafruit RGB backlight negative LCD 20x4 <https://www.adafruit.com/product/498>`_
 
-Compatible with CircuitPython Versions: 2.x
-
 Dependencies
 =============
 This driver depends on:
@@ -42,20 +40,23 @@ The ``Character_LCD`` class interfaces a predefined Character LCD display with C
 
 .. code-block:: python
 
-    import adafruit_character_lcd
+    import board
+    import digitalio
+    import adafruit_character_lcd.character_lcd as character_lcd
 
 You must define the data pins (``RS``, ``EN``, ``D4``, ``D5``, ``D6``, ``D7``) in your code before using the ``Character_LCD`` class.
-If you want to have on/off ``backlight`` functionality, you can also define your backlight as ``lcd_backlight``. Otherwise, the backlight will always remain on. An example of this is below
+If you want to have on/off ``backlight`` functionality, you can also define your backlight as ``lcd_backlight``. Otherwise, the backlight
+will always remain on. The following is an example setup.
 
 .. code-block:: python
 
-    lcd_rs = digitalio.DigitalInOut(D7)
-    lcd_en = digitalio.DigitalInOut(D8)
-    lcd_d7 = digitalio.DigitalInOut(D12)
-    lcd_d6 = digitalio.DigitalInOut(D11)
-    lcd_d5 = digitalio.DigitalInOut(D10)
-    lcd_d4 = digitalio.DigitalInOut(D9)
-    lcd_backlight = digitalio.DigitalInOut(D13)
+    lcd_rs = digitalio.DigitalInOut(board.D7)
+    lcd_en = digitalio.DigitalInOut(board.D8)
+    lcd_d7 = digitalio.DigitalInOut(board.D12)
+    lcd_d6 = digitalio.DigitalInOut(board.D11)
+    lcd_d5 = digitalio.DigitalInOut(board.D10)
+    lcd_d4 = digitalio.DigitalInOut(board.D9)
+    lcd_backlight = digitalio.DigitalInOut(board.D13)
 
 You must also define the size of the CharLCD by specifying its ``lcd_columns`` and ``lcd_rows``:
 
@@ -68,7 +69,7 @@ After you have set up your LCD, we can make the device by calling it
 
 .. code-block:: python
 
-    lcd = adafruit_character_lcd.Character_LCD(lcd_rs, lcd_en, lcd_d4, lcd_d5, lcd_d6, lcd_d7, lcd_columns, lcd_rows, lcd_backlight)
+    lcd = character_lcd.Character_LCD_Mono(lcd_rs, lcd_en, lcd_d4, lcd_d5, lcd_d6, lcd_d7, lcd_columns, lcd_rows, lcd_backlight)
 
 
 To verify that your pins are correct, print a hello message to the CharLCD:
@@ -78,7 +79,7 @@ To verify that your pins are correct, print a hello message to the CharLCD:
     lcd.message = "Hello\nCircuitPython"
 
 
-Custom character example with create_char() is provided within /examples/
+Custom character example with ``create_char()`` is provided within /examples/
 
 
 Contributing
