+++ docs/api.rst
@@ -3,8 +3,12 @@
 
 .. automodule:: adafruit_character_lcd.character_lcd
    :members:
-   :inherited-members:
 
-.. automodule:: adafruit_character_lcd.character_lcd_rgb
+.. automodule:: adafruit_character_lcd.character_lcd_i2c
    :members:
-   :inherited-members:
+
+.. automodule:: adafruit_character_lcd.character_lcd_spi
+   :members:
+
+.. automodule:: adafruit_character_lcd.character_lcd_rgb_i2c
+   :members:
