+++ examples/charlcd_custom_character_nyan_cat.py
@@ -1,10 +1,14 @@
+"""Use custom characters to display Nyan cat"""
 import time
 import board
 import digitalio
-import adafruit_character_lcd.character_lcd_mono as character_lcd
+import adafruit_character_lcd.character_lcd as characterlcd
 
+# Modify this if you have a different sized character LCD
 lcd_columns = 16
 lcd_rows = 2
+
+# Metro M0/M4 Pin Config:
 lcd_rs = digitalio.DigitalInOut(board.D7)
 lcd_en = digitalio.DigitalInOut(board.D8)
 lcd_d7 = digitalio.DigitalInOut(board.D12)
@@ -12,8 +16,10 @@ lcd_d6 = digitalio.DigitalInOut(board.D11)
 lcd_d5 = digitalio.DigitalInOut(board.D10)
 lcd_d4 = digitalio.DigitalInOut(board.D9)
 lcd_backlight = digitalio.DigitalInOut(board.D13)
-lcd = character_lcd.Character_LCD(lcd_rs, lcd_en, lcd_d4, lcd_d5, lcd_d6,
-                                  lcd_d7, lcd_columns, lcd_rows, lcd_backlight)
+
+# Initialise the LCD class
+lcd = characterlcd.Character_LCD_Mono(lcd_rs, lcd_en, lcd_d4, lcd_d5, lcd_d6,
+                                      lcd_d7, lcd_columns, lcd_rows, lcd_backlight)
 
 head = [31, 17, 27, 17, 17, 21, 17, 31]
 
@@ -67,6 +73,8 @@ lcd.clear()
 
 lcd.move_right()
 lcd.message = "\x02\x02\x02\x02\x01\x00\x00\x00\x06\n\x02\x02\x02\x07\x03\x04\x04\x04\x05"
+
+lcd.backlight = True
 
 while True:
     lcd.create_char(4, bot_body2)
