+++ examples/charlcd_customcharacter.py
@@ -1,10 +1,13 @@
-import digitalio
+"""Display a custom character"""
 import board
-import adafruit_character_lcd.character_lcd_mono as character_lcd
+import digitalio
+import adafruit_character_lcd.character_lcd as characterlcd
 
+# Modify this if you have a different sized character LCD
 lcd_columns = 16
 lcd_rows = 2
 
+# Metro M0/M4 Pin Config:
 lcd_rs = digitalio.DigitalInOut(board.D7)
 lcd_en = digitalio.DigitalInOut(board.D8)
 lcd_d7 = digitalio.DigitalInOut(board.D12)
@@ -13,8 +16,9 @@ lcd_d5 = digitalio.DigitalInOut(board.D10)
 lcd_d4 = digitalio.DigitalInOut(board.D9)
 lcd_backlight = digitalio.DigitalInOut(board.D13)
 
-lcd = character_lcd.Character_LCD(lcd_rs, lcd_en, lcd_d4, lcd_d5, lcd_d6,
-                                  lcd_d7, lcd_columns, lcd_rows, lcd_backlight)
+# Initialise the LCD class
+lcd = characterlcd.Character_LCD_Mono(lcd_rs, lcd_en, lcd_d4, lcd_d5, lcd_d6,
+                                      lcd_d7, lcd_columns, lcd_rows, lcd_backlight)
 
 checkmark = bytes([0x0, 0x0, 0x1, 0x3, 0x16, 0x1c, 0x8, 0x0])
 lcd.clear()
