+++ examples/charlcd_i2c_rgb_simpletest.py
@@ -1,8 +1,8 @@
-# Simple test for I2C RGB character LCD shield kit
+"""Simple test for I2C RGB character LCD shield kit"""
 import time
 import board
 import busio
-import adafruit_character_lcd.character_lcd_rgb as character_lcd
+import adafruit_character_lcd.character_lcd_rgb_i2c as character_lcd
 
 # Modify this if you have a different sized Character LCD
 lcd_columns = 16
@@ -12,7 +12,7 @@ lcd_rows = 2
 i2c = busio.I2C(board.SCL, board.SDA)
 
 # Initialise the LCD class
-lcd = character_lcd.Character_LCD_I2C_RGB(i2c, lcd_columns, lcd_rows)
+lcd = character_lcd.Character_LCD_RGB_I2C(i2c, lcd_columns, lcd_rows)
 
 lcd.clear()
 # Set LCD color to red
