+++ examples/charlcd_mono_simpletest.py
@@ -1,8 +1,8 @@
-# Simple test for monochromatic character LCD
+"""Simple test for monochromatic character LCD"""
 import time
 import board
 import digitalio
-import adafruit_character_lcd.character_lcd_mono as character_lcd
+import adafruit_character_lcd.character_lcd as characterlcd
 
 # Modify this if you have a different sized character LCD
 lcd_columns = 16
@@ -18,8 +18,8 @@ lcd_d4 = digitalio.DigitalInOut(board.D9)
 lcd_backlight = digitalio.DigitalInOut(board.D13)
 
 # Initialise the LCD class
-lcd = character_lcd.Character_LCD(lcd_rs, lcd_en, lcd_d4, lcd_d5, lcd_d6,
-                                  lcd_d7, lcd_columns, lcd_rows, lcd_backlight)
+lcd = characterlcd.Character_LCD_Mono(lcd_rs, lcd_en, lcd_d4, lcd_d5, lcd_d6,
+                                      lcd_d7, lcd_columns, lcd_rows, lcd_backlight)
 
 # Turn backlight on
 lcd.backlight = True
@@ -58,6 +58,7 @@ for i in range(len(scroll_msg)):
     lcd.move_left()
 lcd.clear()
 lcd.message = "Going to sleep\nCya later!"
+time.sleep(3)
 # Turn backlight off
 lcd.backlight = False
 time.sleep(2)
